refactor: create proper sub page for move to folder step
